fswatcher_driver
Driver and port to watch the scanned directory for changes and rescan that
directory automatically. Uses FSEvents on Mac, Windows and Linux versions
on the way.

We mirrored the output style of inotifywait so Linux support is easy.

Note, there is a bug, where if the line is > 2048 chars it will skip that
file. This should be fixed as in the unlikley event that a path is that long
(eg. very long track names), that file will never get updated.